fix: added file autocomplete and logic has been transfered from app.tsx
